Refactor of RNNG.
It was difficult to implement word-level search due to the disentanglement
between models, stack, action_history, and token buffer. This meant
that the word level search couldn't easily assume how to create graphs
of actions and tokens. This refactor entangles the structures and models
more to ensure that each model knows how to ensure this graph building.

Some severe bugs were discovered and fixed during the refactor:
* Reduce action would stop popping whenever a non-terminal was found.
  This is incorrect since it also has to check that the non-terminal
  is open. Otherwise reduce will stop prematurely.
* Reduce action would not pop the open non-terminal.