Small bug fixes and enhancements.
Crucially, this commit includes a change where valid actions are not
accounted for during training. Valid actions are still accounted for
during evaluation. The reason for this change is that reducing the
set of actions had an impact on gradient computations. Some gradients,
e.g. compose non-terminals, could become 0 as a result of this.