Split sampling and evaluation into separate scripts.
The reason for doing this is, that samples can be reused across
different evaluations by doing this. That is, for example, relevant
when searching for hyperparameters of generative models. Then there
would be no need search for repeated sampling from a discriminative model,
which is a very time consuming operation.